+++ pkg/engine/coredns/engine_test.go
@@ -106,7 +106,7 @@ func testEngineConfig(listenAddr string, listenPort int) engine.EngineConfig {
 			PrefetchThreshold: 10,
 		},
 		ListenAddr: listenAddr,
-		ListenPort: listenPort,
+		ListenPort: int32(listenPort),
 	}
 }
